requirements.txt
duckdb==1.0.0
Flask==3.0.3
mysql-connector-python==8.1.0
numpy==1.26.0
pandas==2.1.1